# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pesisir Project",
        "description": "API Documentation for Pesisir",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/volunteers": {
            "get": {
                "tags": [
                    "Volunteers"
                ],
                "summary": "Get list of volunteers",
                "operationId": "714cc7e971818196443a1351938ab066",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Volunteers"
                ],
                "summary": "Create new volunteer",
                "operationId": "7d3787a79a518ca185be325f57c863cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "address",
                                    "phone",
                                    "gender",
                                    "reason_desc",
                                    "payment_method",
                                    "act_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "reason_desc": {
                                        "type": "string"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "act_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Volunteer created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/volunteers/{id}": {
            "get": {
                "tags": [
                    "Volunteers"
                ],
                "summary": "Get volunteer by ID",
                "operationId": "4e1b8124c8c58401420b057ba3b6faa8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Volunteer not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Volunteers"
                ],
                "summary": "Delete volunteer by ID",
                "operationId": "2dfca45f6d7982709ab781954f9f9254",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Volunteer deleted successfully"
                    },
                    "404": {
                        "description": "Volunteer not found"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}